Distributed users endpoints into newly created files
New files:
 - booking
 - favorite
 - review
 - store
 - user
